fix: pr-e2e uses pull_request.head.sha instead of pull_request.merge_commit_sha
Reason: pull_request.merge_commit_sha would consistently checkout the second last commit of the pr instead of the current one. This is unintended behaviour. With this change, the setting "Require branches to be up to date before merging" is required to run the test on the git state after the pr would be merged into main by requiring main to be merged into the pr branch before merging back to main.
Source: https://github.com/actions/checkout/issues/518